+++ R/valar.R
@@ -5,7 +5,8 @@
   register <- list(
     valresultat = list(RD = c(2022L, 2026L), RF = c(2022L, 2026L),
                       KF = c(2022L, 2026L)),
-    mandat = list(RD = 2026L, RF = 2026L, KF = 2026L),
+    mandat = list(RD = c(2022L, 2026L), RF = c(2022L, 2026L),
+                  KF = c(2022L, 2026L)),
     kandidaturer = list(RD = 2026L, RF = 2026L, KF = 2026L),
     kandidater = list(RD = 2026L, RF = 2026L, KF = 2026L),
     personroster = list(RD = 2026L, RF = 2026L, KF = 2026L),